refactor: Move thing code to a specific package
Along with the a28de52d519f3f720107258208040bccffafeffb commit
motivation, this patch aims to separate the thing's code base into a
dedicated folder to it.
