+++ pkg/network/msg.go
@@ -1,13 +1,5 @@
 package network
 
-// InMsg represents the message received from the AMQP broker
-type InMsg struct {
-	Exchange   string
-	RoutingKey string
-	Headers    map[string]interface{}
-	Body       []byte
-}
-
 // RegisterRequestMsg is received to register a device
 type RegisterRequestMsg struct {
 	ID   string `json:"id"`
