+++ pkg/thing/delivery/http/thing_proxy.go
@@ -1,4 +1,4 @@
-package network
+package http
 
 import (
 	"bytes"
