+++ pkg/thing/interactors/register_thing_test.go
